requirements.txt
Django==2.2.10
djangorestframework==3.11.2
pytz==2021.1
sqlparse==0.4.1
